+++ src/refract/JSONSchemaVisitor.cc
@@ -18,19 +18,20 @@ namespace refract
     template <typename T>
     void CloneMembers(T* a, const RefractElements* val)
     {
-            for (RefractElements::const_iterator it = val->begin();
-                 it != val->end();
-                 ++it) {
+        for (RefractElements::const_iterator it = val->begin();
+             it != val->end();
+             ++it) {
 
-                if ((*it)->empty()) {
-                    continue;
-                }
-                RenderJSONVisitor v;
-                v.visit(*(*it));
-                IElement *e = v.getOwnership();
-                e->renderType(IElement::rCompact);
-                a->push_back(e);
+            if ((*it)->empty()) {
+                continue;
             }
+
+            RenderJSONVisitor v;
+            v.visit(*(*it));
+            IElement *e = v.getOwnership();
+            e->renderType(IElement::rCompact);
+            a->push_back(e);
+        }
     }
 
 
@@ -201,7 +202,7 @@ namespace refract
             addMember(str->value, renderer.getOwnership());
         }
         else {
-            throw std::logic_error("A property's key in the object is not of type string");
+            throw LogicError("A property's key in the object is not of type string");
         }
 
         if (ext && str) {
@@ -278,6 +279,7 @@ namespace refract
         ObjectElement *o = new ObjectElement;
         ArrayElement::ValueType reqVals;
         std::vector<MemberElement*> varProps;
+        RefractElements oneOfMembers;
 
         o->renderType(IElement::rCompact);
 
@@ -285,38 +287,64 @@ namespace refract
              it != val.end();
              ++it) {
 
-            if (*it) {
-                MemberElement *mr = TypeQueryVisitor::as<MemberElement>(*it);
-                if (!mr) {
-                    continue;
-                }
-
-                if (IsTypeAttribute(*(*it), "required") ||
-                    IsTypeAttribute(*(*it), "fixed") ||
-                    (fixed && !IsTypeAttribute(*(*it), "optional"))) {
-                    StringElement *str = TypeQueryVisitor::as<StringElement>(mr->value.first);
-                    if (str) {
-                        reqVals.push_back(IElement::Create(str->value));
-                    }
-                }
-
-                if (IsVariableProperty(*mr->value.first)) {
-                    varProps.push_back(mr);
-                }
-                else {
-                    JSONSchemaVisitor renderer(pDefs, fixed);
-                    renderer.visit(*(*it));
-                    ObjectElement *o1 = TypeQueryVisitor::as<ObjectElement>(renderer.get());
-                    if (!o1->value.empty()) {
-                        MemberElement *m1 = TypeQueryVisitor::as<MemberElement>(o1->value[0]->clone());
-                        if (m1) {
-                            m1->renderType(IElement::rCompact);
-                            o->push_back(m1);
-                        }
-
-                    }
-                }
+            if (!*it) {
+                continue;
             }
+
+            TypeQueryVisitor type;
+            type.visit(*(*it));
+
+            switch (type.get()) {
+                case TypeQueryVisitor::Member: {
+                    MemberElement *mr = static_cast<MemberElement*>(*it);
+                    if (IsTypeAttribute(*(*it), "required") ||
+                        IsTypeAttribute(*(*it), "fixed") ||
+                        (fixed && !IsTypeAttribute(*(*it), "optional"))) {
+
+                        StringElement *str = TypeQueryVisitor::as<StringElement>(mr->value.first);
+                        if (str) {
+                            reqVals.push_back(IElement::Create(str->value));
+                        }
+                    }
+
+                    if (IsVariableProperty(*mr->value.first)) {
+                        varProps.push_back(mr);
+                    }
+                    else {
+                        JSONSchemaVisitor renderer(pDefs, fixed);
+                        renderer.visit(*(*it));
+                        ObjectElement *o1 = TypeQueryVisitor::as<ObjectElement>(renderer.get());
+                        if (!o1->value.empty()) {
+                            MemberElement *m1 = TypeQueryVisitor::as<MemberElement>(o1->value[0]->clone());
+                            if (m1) {
+                                m1->renderType(IElement::rCompact);
+                                o->push_back(m1);
+                            }
+
+                        }
+                    }
+                }
+                break;
+
+                case TypeQueryVisitor::Select:
+                {
+                    SelectElement* sel = static_cast<SelectElement*>(*it);
+
+                    // FIXME: there is no valid solution for moltiple "SelectElement" in one object.
+
+                    for (SelectElement::ValueType::const_iterator it = sel->value.begin() ; it != sel->value.end() ; ++it) {
+                        JSONSchemaVisitor v(pDefs);
+                        (*it)->content(v);
+                        oneOfMembers.push_back(v.getOwnership());
+                    }
+                }
+                break;
+
+                default:
+                    throw LogicError("Invalid member type of object in MSON definition");
+                    
+            }
+
         }
 
         if (!varProps.empty()) {
@@ -331,10 +359,11 @@ namespace refract
         }
 
         if (!reqVals.empty()) {
-            ArrayElement *a = new ArrayElement;
-            a->renderType(IElement::rCompact);
-            a->set(reqVals);
-            addMember("required", a);
+            addMember("required", new ArrayElement(reqVals, IElement::rCompact));
+        }
+
+        if (!oneOfMembers.empty()) {
+            addMember("oneOf", new ArrayElement(oneOfMembers, IElement::rCompact));
         }
 
         if (fixed) {
@@ -514,6 +543,35 @@ namespace refract
 
         visit(*merged);
         delete merged;
+    }
+
+    void JSONSchemaVisitor::visit(const OptionElement& e)
+    {
+        ObjectElement* props = pObj;
+        RefractElements members;
+        RefractElements oneOfMembers;
+        IncludeMembers(e, members);
+
+        for (OptionElement::ValueType::const_iterator it = members.begin() ; it != members.end() ; ++it) {
+            if (SelectElement* sel = TypeQueryVisitor::as<SelectElement>(*it)) {
+                for (SelectElement::ValueType::const_iterator it = sel->value.begin() ; it != sel->value.end() ; ++it) {
+                    JSONSchemaVisitor v(pDefs);
+                    (*it)->content(v);
+                    oneOfMembers.push_back(v.getOwnership());
+                }
+            }
+            else {
+                visit(*(*it));
+            }
+        }
+
+        pObj = new ObjectElement;
+        pObj->renderType(IElement::rCompact);
+
+        addMember("properties", props);
+        if (!oneOfMembers.empty()) {
+            addMember("oneOf", new ArrayElement(oneOfMembers, IElement::rCompact));
+        }
     }
 
     IElement* JSONSchemaVisitor::get()