+++ src/refract/JSONSchemaVisitor.h
@@ -59,6 +59,8 @@ namespace refract
         void visit(const BooleanElement& e);
         void visit(const ExtendElement& e);
 
+        void visit(const OptionElement& e);
+
         IElement* get();
         IElement* getOwnership();
         std::string getSchema(const IElement& e);